webapi/tct-power-tizen-tests/power/PowerManager_unsetScreenStateChangeListener_successful.html

<!DOCTYPE html>
<!--
Copyright (c) 2012 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        leizhan <zhanx.lei@intel.com>

-->

<html>
<head>
<title>PowerManager_unsetScreenStateChangeListener_successful</title>
<link rel="author" title="Intel" href="http://www.intel.com" />
<link rel="help" href="https://developer.tizen.org/help/topic/org.tizen.web.device.apireference/tizen/power.html" />
<meta name="flags" content="" />
<meta name="assert" content="Check if PowerManager unsetScreenStateChangeListener succeeds" />
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script type="text/javascript">
//==== TEST: PowerManager_unsetScreenStateChangeListener_successful
//==== LABEL Check if unsetScreenStateChangeListener method works correctly
//==== EXECUTION_TYPE manual
//==== STEP Make sure that backlight time is 15 seconds (Settings, Display).
//==== EXPECT (None)
//==== STEP Click 'Run', do not touch the screen, wait until the screen is locked
//==== EXPECT The screen should be locked
//==== STEP Unlock the screen and display the test application
//==== EXPECT The result should be 'Pass'
//==== SPEC Tizen Web API:System:Power:PowerManager:unsetScreenStateChangeListener M
//==== SPEC_URL https://developer.tizen.org/help/topic/org.tizen.web.device.apireference/tizen/power.html
//==== TEST_CRITERIA MNA MNAST
setup({timeout: 90000});

var t = async_test(document.title, {timeout: 90000}), changeDetected = false;
t.step(function () {
    function onScreenStateChanged(previousState, changedState) {
        t.step(function () {
            changeDetected = true;
            assert_unreached("OnScreenStateChanged callback shouldn't " +
                "be invoked!");
        });
    }

    tizen.power.release("SCREEN");
    tizen.power.release("CPU");
    tizen.power.setScreenStateChangeListener(onScreenStateChanged);
    tizen.power.unsetScreenStateChangeListener();
    tizen.power.turnScreenOn();
    setTimeout(function () {
        t.step(function () {
            assert_false(changeDetected, "screen state change shouldn't be " +
                "detected.");
        });
        t.done();
    }, 20000);
});

</script>
</body>
</html>
